+++ rouver/handler.py
@@ -6,7 +6,7 @@ from http import HTTPStatus
 from json import JSONDecodeError, loads as json_loads
 from typing import Any, Protocol, cast
 from urllib.parse import unquote
- 
+
 from werkzeug.exceptions import UnsupportedMediaType
 from werkzeug.wrappers import Request
 
@@ -25,8 +25,7 @@ from rouver.types import Header, StartResponse, WSGIEnvironment
 
 
 class _Closeable(Protocol):
-    def close(self) -> object:
-        ...
+    def close(self) -> object: ...
 
 
 class RouteHandlerBase(collections.abc.Iterable):
@@ -50,9 +49,7 @@ class RouteHandlerBase(collections.abc.Iterable):
     ...         ])
     """
 
-    def __init__(
-        self, environ: WSGIEnvironment, start_response: StartResponse
-    ) -> None:
+    def __init__(self, environ: WSGIEnvironment, start_response: StartResponse) -> None:
         self.request = Request(environ)
         self.start_response = start_response
         self._argument_parser: ArgumentParser | None = None
@@ -185,9 +182,7 @@ class RouteHandlerBase(collections.abc.Iterable):
     def created_as_json(
         self, url_part: str, json: str | bytes | Any
     ) -> Iterable[bytes]:
-        return created_as_json(
-            self.request, self.start_response, url_part, json
-        )
+        return created_as_json(self.request, self.start_response, url_part, json)
 
     def temporary_redirect(self, url_part: str) -> Iterable[bytes]:
         return temporary_redirect(self.request, self.start_response, url_part)